ruyaa-takhless-bot/requirements.txt

aiogram==3.*
aiosqlite
python-dotenv
fastapi
redis>=5
